Got Main Menu Image Up
Added in a rough draft for the main menu and briefly enabled it (crappily) in the UI to see if I could get it to display. It worked about as well as intended. Next step will thus be to strip away the old xaml UI until I can transition into my idea of a menu.