Core Navigation Framework Started
I felt like committing even though this part isn't really close to done because I was excited that I put in something new without a real hitch. At this stage, I should just be able to swap old code around (or just re-enable a little bit of it) and be well on my way to a more full featured navigation system.