No Idea Right Now.
It seems totally confusing to me that I can play a sound effect through a bit of Ammo, but not directly. I'll need to keep digging to figure out what the hell is going on here. I SHOULD be able to just play audio directly from a sound effect, but I have no idea why it's breaking.