Idea found.
I discovered that the issue I was encountering was sourced from failing to declare the sound effect object I wished to use in a header file. Apparently, I need to declare it in a header file or I get some sort of odd loading/memory issue that comes across as a totally unhelpful bug.