Consolidated "About Me", Added Settings
I wanted to include the option for custom keybindings. It turns out, the original sample code is doing things a bit strangely by relying on Virtual Keys instead of actual key presses with the ability to suggest reasonable defaults and change them later. My particular issue at the moment is that in my layout, the W key corresponds to ,/<. This isn't a virtual key I should be using and I shouldn't be using these anyhow. I just want to react to hardware events instead of translated ones if possible. We'll see what tomorrow turns up. I probably have to rework the basic input scheme that was provided in the sample.