New Keybinding Array In Use
Got the mapping down and proceeded to add a bit of code to utilize it. Basically, the original functionality of the program is restored and nothing is really different, but this paves the way to change the functionality such that functions can be rebound to different keys. The next step is to capture input from the keyboard on the options screen and then both display the new binding and map it up to be used as an actual bind. Not sure if I can still use a textbox in order to make it work. Further investigation required.