Key Bindings Complete (Mostly)
It is now possible to rebind keys to pretty much anything on the keyboard. The only sort of unfortunate/strange things left to potentially address are the current inability to bind mouse presses and a slight awkwardness in exiting the binding box since I didn't implement an automatic focus transfer for the textbox after the key down event.