Rectangular Prism Added
Added a temporary (incorrect and incomplete) version of a rectangular prism class to the sample. It is currently unused. Eventually, I will use it to recast the game as a platformer. I think that before I do that, I want to rework the collision checking to use a version of the line sweep algorithm, but I will need to both explore that a bit more personally and change the current code to support it.