Axis Overlap Detection Algorithm Complete
I'm confident at this point that I created a solid algorithm to dected overlapping line segments in an efficient manner that ought to scale well with large object numbers. Next steps include adding support to objects to take advantage of this algorithm and ultimately replacing the old collision detection system with this new tech.