O M E G A L U L (Fixed Slowness)
It turns out that the [xyz]AxisSweep lists were never getting cleared...so of course the sort was taking a long time. lol. The lists at least doubled in size every collision detection test. Now it runs smoothly again. NB: The algorithm can still be agressively improved and the sorting test to test can definitely be dropped in favor of a more sophisticated system.