Pretty Pictures & Little Things
Added in some fancier graphics and did a bit of cleaning up. Additionally, made some progress in figuring out professional (ish) collision stuff. Hopefully the next completed thing will be the collision as a whole. It is tough to process whether or not it is working at this time though. Need work to change up the world format and space to get it nailed.