Struggle Bus Docked At The Station
It was quite the journey to get this cylinder bounding box figured out. More geometry and misfires in my brain than I care to admit occurred. I'm not sure I consented to this when I decided to be a programmer. LuL. Jk. I'm a nerd at heart. This math is fine. This note is maybe not though. Let's make it a bit more professional. I'm relieved to finally be done slamming my forehead against a wall in regards to figuring out the bounding box of cylinders (Note: The current version is actually pretty rough. It gets particularly bad when non-axis aligned cylinders enter the picture, but such is life for now.) At this point, all that really remains is to get it ready to be shipped out. I think precision collision can be dealt with as I go along, but I think with this change made I should have a decent enough starting point to start trying to pack it up for my resume.